+++ rpibuild.sh
@@ -1,1 +1,0 @@
-cc bin/snob.o -o bin/snob  -rdynamic -L/usr/bin/../lib/crystal/lib -lyaml -lpcre -lm -lgc -lpthread /usr/share/crystal/src/ext/libcrystal.a -levent -lrt -ldl