+++ scripts/mkdocs.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+asciidoctor -r asciidoctor-epub3 -b epub3 -D docs README.adoc
+asciidoctor -r asciidoctor-pdf -b pdf -D docs README.adoc