+++ scripts/rpibuild64.sh
@@ -1,0 +1,1 @@
+cc bin/snob-aarch64.o -o bin/snob  -rdynamic -L/usr/bin/../lib/crystal/lib -lyaml -lpcre -lm -lgc -lpthread -levent -lrt -ldl